fix(aapd-1524): fix migrations, remove log

+++ packages/database/src/migrations/20240509091633_site_area_to_decimal/migration.sql
@@ -3,7 +3,7 @@ BEGIN TRY
 BEGIN TRAN;
 
 -- AlterTable
-ALTER TABLE [dbo].[AppellantSubmission] ALTER COLUMN [siteAreaSquareMetres] FLOAT(53) NULL;
+ALTER TABLE [dbo].[AppellantSubmission] ALTER COLUMN [siteAreaSquareMetres] DECIMAL(32,16) NULL;
 
 COMMIT TRAN;
